# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/attendances:
    get:
      tags:
        - Attendances
      summary: List Attendances
      description: Get a paginated list of Attendances with optional filtering and sorting
      operationId: listAttendances
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Attendances
      summary: Create Attendance
      description: Create a new Attendance
      operationId: createAttendance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAttendanceRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Attendance'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/attendance_clocks:
    get:
      tags:
        - AttendanceClocks
      summary: List AttendanceClocks
      description: Get a paginated list of AttendanceClocks with optional filtering and sorting
      operationId: listAttendanceClocks
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceClockList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - AttendanceClocks
      summary: Create AttendanceClock
      description: Create a new AttendanceClock
      operationId: createAttendanceClock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAttendanceClockRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceClock'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    Attendance:
      type: object
      required:
        - id
        - company_id
        - employee_id
        - date
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        date:
          type: string
          format: date
        schedule:
          nullable: true
          type: object
        clockin:
          nullable: true
          type: string
          format: time
        clockout:
          nullable: true
          type: string
          format: time
        time_debt:
          nullable: true
          type: object
        timeoff:
          nullable: true
          type: object
        metadata:
          type: object
...
    CreateAttendanceRequest:
      type: object
      required:
        - company_id
        - employee_id
        - date
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        date:
          type: string
          format: date
        schedule:
          nullable: true
          type: object
        clockin:
          nullable: true
          type: string
          format: time
        clockout:
          nullable: true
          type: string
          format: time
        time_debt:
          nullable: true
          type: object
        timeoff:
          nullable: true
          type: object
        metadata:
          type: object
...
    AttendanceList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Attendance'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    AttendanceClock:
      type: object
      required:
        - id
        - company_id
        - session_id
        - employee_id
        - date
        - punched_at
        - direction
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        session_id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        date:
          type: string
          format: date
        punched_at:
          type: string
          format: date-time
        direction:
          type: string
        metadata:
          type: object
...
    CreateAttendanceClockRequest:
      type: object
      required:
        - company_id
        - session_id
        - employee_id
        - date
        - punched_at
        - direction
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        session_id:
          type: string
          format: uuid
        employee_id:
          type: string
          format: uuid
        date:
          type: string
          format: date
        punched_at:
          type: string
          format: date-time
        direction:
          type: string
        metadata:
          type: object
...
    AttendanceClockList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AttendanceClock'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []